Se creo una clase Menu para administrar mejor los menus y se dejo libre el método main del sistema, sólo para instanciar la clase de Menu
